Exctract links from Navbar into their own component
