+++ src/components/navbar/Navbar.tsx
@@ -1,0 +1,21 @@
+import React, { FunctionComponent } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Link } from 'react-router-dom';
+import { NavLinks } from './NavLinks';
+
+export const Navbar: FunctionComponent = () => {
+  const { t } = useTranslation();
+  return (
+    <nav className="flex flex-row bg-brand items-center">
+      <Link to="/" aria-label={t('routes.root')}>
+        <img
+          src="/nav-logo.svg"
+          style={{ height: 40, marginLeft: 29 }}
+          alt=""
+          className="pr-5"
+        />
+      </Link>
+      <NavLinks />
+    </nav>
+  );
+};
